05_FastAPI_DL_Models_Practice/pyproject.toml
[project]
name = "fastapi-dl"
version = "0.1.0"
description = ""
authors = [
    {name = "Hassan Mehmood", email = "ihassan463m@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]>=0.115.12,<0.116.0",
    "sqlalchemy>=2.0.40,<3.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "torch>=2.6.0,<3.0.0",
    "torchvision>=0.21.0,<0.22.0",
    "tensorflow-cpu>=2.11.0,<3.0.0",
    "chainlit==2.4.400"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
